Added first buff, buff system, shared cooldown component, primitive logic for AI

+++ BardSimV2/Auxiliars/Cooldown.cs
@@ -9,12 +9,12 @@ namespace BardSimV2
     class Cooldown
     {
         private decimal baseRecast;
-        private decimal start;
+        private ulong start;
 
         public decimal BaseRecast { get => baseRecast; }
-        public decimal Start { get => start; set => start = value; }
+        public ulong Start { get => start; set => start = value; }
 
-        public Cooldown(decimal baseRecast, decimal start = 0)
+        public Cooldown(decimal baseRecast, ulong start = 0)
         {
             this.baseRecast = baseRecast;
             this.start = start;
